Increase prominence of Resin Base Images Wiki
Previously this was not so easily accessible, these changes rename the page to
make use of the phrase 'base image' which is the more commonly known name for
images, add a link to the page on the references tab and add a blurb about base
images with a link to the introduction page.

Solves issue #248.