add category header pages
Some of the categories are harder to understand, when clicking on the header
in the navigation bar takes the user right to the first item in the category.
It makes the other items in the category easy to miss in our navigation bar.
Add some category pages, from where the rest of the items are linked.