Fix the improve this doc button on dynamic pages
Previously, the "Improve this doc" button failed for pages that are generated dynamically, as the page does not exist anywhere as a single markdown file that can be edited. Some logic has been added so that any dynamic pages just link to the shared folder, which contains any markdown files that would need to be edited
to fix the page.

In addition, the getting started page links have been reordered to make more sense, for example "getting-started/raspberrypi3/nodejs" instead of "raspberrypi3/nodejs/getting-started". The redirects have been updated to handle links in the old format.

Change-type: patch