Updates for supported devices
The supported devices page was displaying anchor links for the headings that did not work, as the headings are not generated via markdown and therefore the elements don't get an ID automatically assigned. The code has been updated to not attach anchor links to elements without IDs.

The table of device names was also updated to remove the parallella device type, as it is discontinued. A small typo was also fixed.

Change-type: patch