fix: Update navigation to reflect new base image page name
The navigation needs to be updated to reflect that the base image page name is no longer 'resin-base-image'. This was changed as part of the effort to make the docs more name agnostic. A small change to the FAQ page is also included to account for removing 'resin' from a heading.

Change-type: patch