advanced: Improve docs on modifying config.txt
We fix the reference to config.txt having to be modified
locally after first boot  - it's actually the other way round,
since the supervisor will apply whatever the API says it
should after first boot (and only what was there before
first boot will be kept).

We also expand on how to set dtoverlay and dtparam values,
with the new way to set several values as an array.

Closes #882

Change-type: patch
Signed-off-by: Pablo Carranza Velez <pablo@resin.io>