Revert using S3 images for supported device types
Earlier I didn't run it in watch mode and I thought all of them work :/
It turns out we still don't have logos in S3 for all device types, and
half-baked solution will only make it more messy than it already is.
Once we resolve the issue on where the images are stored and we have fully
migrated to it, I'll revisit this.

Change-type: patch
Signed-off-by: Stevche Radevski <stevche@balena.io>