Be more explicit about risks of using advanced configuration variables
This came from a particular set of changes to config.txt file
on a Raspberry Pi making the device not boot but the problem is general
enough to apply to all devices. There are many ways of misconfiguring
the variables to make the system not bootable, certainly more that we
would be able to list in the docs, therefore the suggestion is to try
in a controlled environment before rolling to production.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>